ID3v2: Move Popularimeter parsing to `Popularimeter::from_bytes`
